requirements.txt
Flask
openai
python-docx
requests
Pillow
pytesseract
pdf2image
pdfplumber
python-dotenv
